feat: add embedding kernel and benchmarks (for multiples of 32 tpr)

+++ include/llmetal/cpu/embedding.hpp
@@ -9,7 +9,7 @@ void embedding(
     llmetal::CpuTensor<float> &table,   // [vocab_size, hidden_size]
     llmetal::CpuTensor<std::uint32_t> &ids, // [batch_size, sequence_length]
     llmetal::CpuTensor<float> &output,   // [batch_size, sequence_length, hidden_size]
-    std::size_t vocab_size
+    std::uint32_t vocab_size
 );
 
 } // namespace llmetal::cpu
